Fix bug where the mod directory would not be created

+++ OpenKh.Patcher/PatcherProcessor.cs
@@ -59,7 +59,8 @@ namespace OpenKh.Patcher
 
         public void Patch(string originalAssets, string outputDir, Metadata metadata, string modBasePath)
         {
-            Directory.Delete(outputDir, true);
+            if (Directory.Exists(outputDir))
+                Directory.Delete(outputDir, true);
             Directory.CreateDirectory(outputDir);
 
             var context = new Context(metadata, originalAssets, modBasePath, outputDir);
